+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 - New API: deformetrica can now be instantiated from python
 - Automatic dimension detection from input file
+- Dataset file no longer mandatory. fixes #6
 
 
 ## [4.0.0] - 2018-06-14